jdbcTemplate 적용 - 추상화(다른 DB 고려)
