+++ pom.xml
@@ -25,7 +25,7 @@
 		<dependency>
 			<groupId>org.springframework</groupId>
 			<artifactId>spring-jdbc</artifactId>
-			<version>3.0.7.RELEASE</version>
+			<version>5.3.19</version>
 		</dependency>
 		<dependency>
 			<groupId>mysql</groupId>
