+++ src/main/java/springbook/user/dao/DaoFactory.java
@@ -12,8 +12,8 @@ import javax.sql.DataSource;
 @Configuration
 public class DaoFactory {
     @Bean
-    public UserDao userDao(){
-        UserDao userDao = new UserDao();
+    public UserDaoJdbc userDao(){
+        UserDaoJdbc userDao = new UserDaoJdbc();
         userDao.setDataSource(dataSource());
         return userDao;
     }
