+++ src/main/java/springbook/user/dao/UserDao.java
@@ -1,66 +1,17 @@
 package springbook.user.dao;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
-
-import org.springframework.dao.DataAccessException;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
-import org.springframework.jdbc.core.ResultSetExtractor;
-import org.springframework.jdbc.core.RowMapper;
-import springbook.user.dao.context.JdbcContext;
-import springbook.user.dao.strategy.StatementStrategy;
 import springbook.user.domain.User;
 
-import javax.sql.DataSource;
+import java.util.List;
 
-public class UserDao {
-    
-    private JdbcTemplate jdbcTemplate;
-
-    public void setDataSource(DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);        
-    }
-    
-    private RowMapper<User> userMapper = new RowMapper<User>() {
-        @Override
-        public User mapRow(ResultSet resultSet, int i) throws SQLException {
-            User user = new User();
-            user.setId(resultSet.getString("id"));
-            user.setName(resultSet.getString("name"));
-            user.setPassword(resultSet.getString("password"));
-            return user;
-        }
-    };
-    
-    public List<User> getAll(){
-        return this.jdbcTemplate.query("select * from users order by id", this.userMapper);
-    }
-    
-    public User get(String id) {
-        return this.jdbcTemplate.queryForObject("select * from users where id = ?",
-                new Object[]{id}, this.userMapper);
-    }
-    
-    public void delete(String id) {
-        this.jdbcTemplate.update("delete from users where id = ?", id);
-    }
-
-    public void add(final User user) {
-        this.jdbcTemplate.update("insert into users(id, name, password) values(?, ?, ?)",
-                user.getId(), user.getName(), user.getPassword());
-    }
-    
-    public void deleteAll() {
-        this.jdbcTemplate.update("delete from users");
-    }
-    
-    public int getCount() {
-        return this.jdbcTemplate.queryForInt("select count(*) from users");
-    }
-
+/**
+ * @author user
+ */
+public interface UserDao {
+    public List<User> getAll();
+    public User get(String id);
+    public void delete(String id);
+    public void add(User user);
+    public void deleteAll();
+    public int getCount();
 }
