+++ src/main/resources/applicationContext.xml
@@ -13,7 +13,7 @@
     </bean>
     
     <bean id="userDao"
-          class="springbook.user.dao.UserDao">
+          class="springbook.user.dao.UserDaoJdbc">
         <property name="dataSource" ref="dataSource" />
     </bean>
 </beans>
