+++ src/main/resources/test-applicationContext.xml
@@ -5,7 +5,7 @@
         http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
     
     <bean id="userDao"
-          class="springbook.user.dao.UserDao">
+          class="springbook.user.dao.UserDaoJdbc">
         <property name="dataSource" ref="dataSource" />
     </bean>
